pr: moved TrackingRequest creation
