+++ shared/src/commonMain/kotlin/io/customer/shared/tracking/api/model/Extensions.kt
@@ -1,10 +1,7 @@
 package io.customer.shared.tracking.api.model
 
-import io.customer.shared.sdk.meta.IdentityType
 import io.customer.shared.tracking.constant.QueueTaskStatus
-import io.customer.shared.tracking.model.Activity
 import io.customer.shared.tracking.model.TrackingError
-import io.customer.shared.tracking.model.type
 
 internal val TrackingError.taskStatus: QueueTaskStatus
     get() = if (reason.isFixable) QueueTaskStatus.FAILED
@@ -23,71 +20,3 @@ internal val BatchTrackingResponse.isSuccessful: Boolean
 
 internal val BatchTrackingResponse.isServerUnavailable: Boolean
     get() = statusCode >= 500
-
-internal fun Activity.toTrackingRequest(
-    identityType: IdentityType,
-    profileIdentifier: String,
-) = when (this) {
-    is Activity.IdentifyProfile -> TrackingRequest(
-        identityType = identityType,
-        profileIdentifier = profileIdentifier,
-    )
-    is Activity.AddDevice -> TrackingRequest(
-        identityType = identityType,
-        profileIdentifier = profileIdentifier,
-        device = device,
-    )
-    is Activity.DeleteDevice -> TrackingRequest(
-        identityType = identityType,
-        profileIdentifier = profileIdentifier,
-        device = device,
-    )
-    is Activity.Metric -> TrackingRequest(
-        identityType = identityType,
-        profileIdentifier = profileIdentifier,
-        metric = Metric(
-            deliveryID = deliveryId,
-            deviceToken = deviceToken,
-            event = metricEvent,
-            timestamp = timestamp,
-        ),
-    )
-    is Activity.Page -> TrackingRequest(
-        identityType = identityType,
-        profileIdentifier = profileIdentifier,
-        name = name,
-    )
-    is Activity.Screen -> TrackingRequest(
-        identityType = identityType,
-        profileIdentifier = profileIdentifier,
-        name = name,
-    )
-    is Activity.Event -> TrackingRequest(
-        identityType = identityType,
-        profileIdentifier = profileIdentifier,
-        name = name,
-    )
-}
-
-private val IdentityType.key: String
-    get() = when (this) {
-        IdentityType.CIO_ID -> "cio_id"
-        IdentityType.ID -> "id"
-        IdentityType.EMAIL -> "email"
-    }
-
-private fun Activity.TrackingRequest(
-    identityType: IdentityType,
-    profileIdentifier: String,
-    name: String? = null,
-    device: Device? = null,
-    metric: Metric? = null,
-) = TrackingRequest(
-    type = type,
-    timestamp = timestamp,
-    identifiers = mapOf(identityType.key to profileIdentifier),
-    attributes = attributes,
-    name = name,
-    device = device,
-    metric = metric,
-)
